\name{FunctQuant}
\alias{FunctQuant}
\docType{package}
\title{
Functional Quantization
}
\description{
\packageDescription{FunctQuant}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FunctQuant}
\packageIndices{FunctQuant}
}
\author{
Maintainer: \packageMaintainer{FunctQuant}
}

~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
\code{\link{GFQ}}, \code{\link{CVT}}, \code{\link{StochGradient}} and \code{\link{Karhunen_Loeve}}
}